requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.103.0
requests==2.31.0
numpy==1.24.3
scikit-learn==1.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
